Move `registerChatModelChangeListeners` onto chat service
No sure what this method was doing on the sessions service. Really should be deleted entirely but moving to a better home on the chat service as a first step
