+++ src/vs/workbench/contrib/chat/browser/agentSessions/localAgentSessionsController.ts
@@ -57,7 +57,7 @@ export class LocalAgentsSessionsController extends Disposable implements IChatSe
 			this._onDidChangeChatSessionItems.fire();
 		};
 
-		this._register(this.chatSessionsService.registerChatModelChangeListeners(this.chatService, Schemas.vscodeLocalChatSession, refreshItems));
+		this._register(this.chatService.registerChatModelChangeListeners(Schemas.vscodeLocalChatSession, refreshItems));
 
 		this._register(this.chatService.onDidDisposeSession(e => {
 			const session = e.sessionResource.filter(resource => getChatSessionType(resource) === this.chatSessionType);
